fastapi-project/requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-dotenv
